+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.37] - 2019-02-25
+### Changed
+- Enable select2 in conditional-forms when `data-widget="select2"` is set
+- Improved SpacingType form template
+
 ## [1.4.36] - 2019-02-22
 ### Added
 - Added a form attribute to enable conditional forms in page form, just add a `data-condition` attribute to the form element to trigger a reload on change. This can also be used in element edit modals.